feat: 🎸 grpc ui upp and running
