+++ rest-ui/package.json
@@ -23,7 +23,6 @@
     "@loadable/babel-plugin": "^5.13.2",
     "babel-eslint": "^10.0.2",
     "babel-loader": "^8.0.6",
-    "classnames": "^2.2.6",
     "clean-webpack-plugin": "^3.0.0",
     "compression-webpack-plugin": "^6.0.2",
     "copy-webpack-plugin": "^6.2.0",
