+++ rest-ui/src/main/js/components/grpc/highlight-code.js
@@ -1,25 +1,16 @@
 import { Light as SyntaxHighlighter } from 'react-syntax-highlighter'
 import json from 'react-syntax-highlighter/dist/esm/languages/hljs/json'
+import xml from 'react-syntax-highlighter/dist/esm/languages/hljs/xml'
 import obsidian from 'react-syntax-highlighter/dist/esm/styles/hljs/obsidian'
-import React, { Component } from 'react'
-
+import React from 'react'
+import { stringify } from './utils'
 SyntaxHighlighter.registerLanguage('json', json)
+SyntaxHighlighter.registerLanguage('xml', xml)
 
-export default class HighlightCode extends Component {
-  render () {
-    let { className, code } = this.props
-    if (typeof code !== 'string') {
-      code = JSON.stringify(code, null, 2)
-    }
-    return (
-      <div className="highlight-code">
-        <SyntaxHighlighter
-          className={className + ' microlight'}
-          style={obsidian}
-        >
-          {code}
-        </SyntaxHighlighter>
-      </div>
-    )
-  }
-}
+export const HighlightCode = ({ className, code }) => (
+  <div className="highlight-code">
+    <SyntaxHighlighter className={className + ' microlight'} style={obsidian}>
+      {stringify(code)}
+    </SyntaxHighlighter>
+  </div>
+)
